# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "RunDetail": {
        "properties": {
          "baseline": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline"
          },
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfigDoc"
              },
              {
                "type": "null"
              }
            ]
          },
          "ref": {
            "title": "Ref",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StatusDoc"
              },
              {
                "type": "null"
              }
            ]
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          }
        },
        "required": [
          "ref",
          "config",
          "status",
          "summary"
        ],
        "title": "RunDetail",
        "type": "object"
      },
      "RunEntry": {
        "description": "One line of the run table.",
        "properties": {
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "max_steps": {
            "title": "Max Steps",
            "type": "integer"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "ref": {
            "title": "Ref",
            "type": "string"
          },
          "stale": {
            "default": false,
            "title": "Stale",
            "type": "boolean"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "step": {
            "title": "Step",
            "type": "integer"
          },
          "updated": {
            "title": "Updated",
            "type": "number"
          },
          "val_bpb": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Val Bpb"
          }
        },
        "required": [
          "ref",
          "state",
          "step",
          "max_steps",
          "val_bpb",
          "updated"
        ],
        "title": "RunEntry",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "nanoscope",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/runs": {
      "get": {
        "description": "Every run under a ref prefix (all when empty), optionally only those in one state.",
        "operationId": "list_runs_api_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "prefix",
            "required": false,
            "schema": {
              "default": "",
              "title": "Prefix",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunEntry"
                  },
                  "title": "Response List Runs Api Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request cannot be done as asked"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An error (RFC 9457)"
          }
        },
        "summary": "List Runs",
        "tags": [
          "runs"
        ]
      },
      "post": {
        "description": "Queue a run on the interactive lane. A run that is already done returns it (200); one\nalready queued or running returns its job (202). The model is read, never imported here:\nthe worker runs it.",
        "operationId": "submit_run_api_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Submit Run Api Runs Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request cannot be done as asked"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An error (RFC 9457)"
          }
        },
        "summary": "Submit Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{ref}": {
      "get": {
        "description": "A run's config, status and a summary of its metrics.",
        "operationId": "get_run_api_runs__ref__get",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request cannot be done as asked"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An error (RFC 9457)"
          }
        },
        "summary": "Get Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{ref}/blockstats": {
      "get": {
        "description": "Per-block statistics recorded at each eval step (runs started with block_stats=True).",
        "operationId": "run_blockstats_api_runs__ref__blockstats_get",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Run Blockstats Api Runs  Ref  Blockstats Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request cannot be done as asked"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An error (RFC 9457)"
          }
        },
        "summary": "Run Blockstats",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{ref}/checkpoints": {
      "get": {
        "description": "The checkpoint files of a run (the weights are not served).",
        "operationId": "run_checkpoints_api_runs__ref__checkpoints_get",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CheckpointEntry"
                  },
                  "title": "Response Run Checkpoints Api Runs  Ref  Checkpoints Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request cannot be done as asked"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An error (RFC 9457)"
          }
        },
        "summary": "Run Checkpoints",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/api/runs/{ref}/generate": {
      "post": {
        "description": "Text from a trained run. It is a `generate` job on the interactive lane: a worker keeps\nrecently used models loaded and does the work, never this process. The answer comes back\nin this response; if the generation outruns `timeout` it is stopped (504), and with no\nworker to take the job it is cancelled (503).",
        "operationId": "generate_text_api_runs__ref__generate_post",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Generate Text Api Runs  Ref  Generate Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request cannot be done as asked"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An error (RFC 9457)"
          }
        },
        "summary": "Generate Text",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/api/runs/{ref}/resume": {
      "post": {
        "description": "Continue a stopped or interrupted run from its last checkpoint, with exactly the\nconfiguration it started with. A finished run has nothing to resume (409).",
        "operationId": "resume_run_api_runs__ref__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Resume Run Api Runs  Ref  Resume Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request cannot be done as asked"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An error (RFC 9457)"
          }
        },
        "summary": "Resume Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{ref}/samples": {
      "get": {
        "description": "The generated samples, one per `sample_interval`, oldest first.",
        "operationId": "run_samples_api_runs__ref__samples_get",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Sample"
                  },
                  "title": "Response Run Samples Api Runs  Ref  Samples Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request cannot be done as asked"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An error (RFC 9457)"
          }
        },
        "summary": "Run Samples",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{ref}/stop": {
      "post": {
        "description": "Ask a run to stop at its next step (it saves a checkpoint, and `resume` continues it);\na job that has not started is cancelled. Works for runs started outside the API too.",
        "operationId": "stop_run_api_runs__ref__stop_post",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/nanoscope__server__routes__runs__StopResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request cannot be done as asked"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An error (RFC 9457)"
          }
        },
        "summary": "Stop Run",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/api/studies/{name}/run": {
      "post": {
        "description": "Queue a study (explore mode). A worker loads the spec, which imports its models, and puts\nevery unfinished run on the batch lane; watch them with `/jobs` or `/events?prefix=`.",
        "operationId": "run_study_api_studies__name__run_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request cannot be done as asked"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An error (RFC 9457)"
          }
        },
        "summary": "Run Study",
        "tags": [
          "studies"
        ]
      }
    },
...
    "/api/validate/run": {
      "post": {
        "description": "Every problem with a run request at once, locked blocks included: what `run()` would\nrefuse, without running anything or importing the model.",
        "operationId": "validate_run_api_validate_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunValidation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Validation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request cannot be done as asked"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An error (RFC 9457)"
          }
        },
        "summary": "Validate Run",
        "tags": [
          "validate"
        ]
      }
    },
...
  }
}